Use functions instead of classes in tile rendering
